refactored code so logic to create a base64 image could be recycled
